Opportunistic fix package for error prone artifacts
This has been bothering me, and moving while we're at it. Not breaking since it's expected to be pick up via ServiceLoader and not a dependency
