+++ static-analysis/autodispose-error-prone-checker/src/test/resources/com/uber/autodispose/errorprone/UseAutoDisposeDefaultClassPositiveCases.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.uber.autodispose.error.prone.checker;
+package com.uber.autodispose.errorprone;
 
-import com.uber.autodispose.AutoDispose;
+import com.uber.autodispose.errorprone.UseAutoDispose;
 import com.uber.autodispose.lifecycle.CorrespondingEventsFunction;
 import com.uber.autodispose.lifecycle.LifecycleEndedException;
 import com.uber.autodispose.lifecycle.LifecycleScopeProvider;
@@ -33,12 +33,11 @@ import io.reactivex.annotations.Nullable;
 import io.reactivex.subjects.BehaviorSubject;
 import org.reactivestreams.Subscriber;
 
-import static com.uber.autodispose.AutoDispose.autoDisposable;
-
 /**
- * Cases that use {@link AutoDispose} and should not fail the {@link UseAutoDispose} check.
+ * Cases that don't use autodispose and should fail the {@link UseAutoDispose} check.
  */
-public class UseAutoDisposeNegativeCases implements LifecycleScopeProvider<TestLifecycleScopeProvider.TestLifecycle> {
+public class UseAutoDisposeDefaultClassPositiveCases
+    implements LifecycleScopeProvider<TestLifecycleScopeProvider.TestLifecycle> {
 
   private final BehaviorSubject<TestLifecycleScopeProvider.TestLifecycle> lifecycleSubject = BehaviorSubject.create();
 
@@ -80,41 +79,41 @@ public class UseAutoDisposeNegativeCases implements LifecycleScopeProvider<TestL
     return LifecycleScopes.resolveScopeFromLifecycle(this);
   }
 
-  public void observable_subscribeWithAutoDispose() {
-    Observable.just(1)
-        .as(autoDisposable(this))
+  public void observable_subscribeWithoutAutoDispose() {
+    Observable.empty()
+        // BUG: Diagnostic contains: Always apply an AutoDispose scope before subscribing within defined scoped elements.
         .subscribe();
   }
 
-  public void single_subscribeWithAutoDispose() {
+  public void single_subscribeWithoutAutoDispose() {
     Single.just(true)
-        .as(autoDisposable(this))
+        // BUG: Diagnostic contains: Always apply an AutoDispose scope before subscribing within defined scoped elements.
         .subscribe();
   }
 
-  public void completable_subscribeWithAutoDispose() {
+  public void completable_subscribeWithoutAutoDispose() {
     Completable.complete()
-        .as(autoDisposable(this))
+        // BUG: Diagnostic contains: Always apply an AutoDispose scope before subscribing within defined scoped elements.
         .subscribe();
   }
 
-  public void maybe_subscribeWithAutoDispose() {
-    Maybe.just(1)
-        .as(autoDisposable(this))
+  public void maybe_subscribeWithoutAutoDispose() {
+    Maybe.empty()
+        // BUG: Diagnostic contains: Always apply an AutoDispose scope before subscribing within defined scoped elements.
         .subscribe();
   }
 
-  public void flowable_subscribeWithAutoDispose() {
-    Flowable.just(1)
-        .as(autoDisposable(this))
+  public void flowable_subscribeWithoutAutoDispose() {
+    Flowable.empty()
+        // BUG: Diagnostic contains: Always apply an AutoDispose scope before subscribing within defined scoped elements.
         .subscribe();
   }
 
-  public void parallelFlowable_subscribeWithAutoDispose() {
+  public void parallelFlowable_subscribeWithoutAutoDispose() {
     Subscriber<Integer>[] subscribers = new Subscriber[] {};
     Flowable.just(1, 2)
         .parallel(2)
-        .as(autoDisposable(this))
+        // BUG: Diagnostic contains: Always apply an AutoDispose scope before subscribing within defined scoped elements.
         .subscribe(subscribers);
   }
 }
